% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult.R
\docType{data}
\name{adult}
\alias{adult}
\title{2018 Census data from the UCI repository.}
\format{
A data frame with 48842 rows and 14 variables:
\describe{
\item{age}{continuous}
\item{workclass}{Private, Self-emp-not-inc, Self-emp-inc, Federal-gov, Local-gov, State-gov, Without-pay, Never-worked.}
\item{fnlwgt}{continuous}
\item{education}{Bachelors, Some-college, 11th, HS-grad, Prof-school, Assoc-acdm, Assoc-voc, 9th, 7th-8th, 12th, Masters, 1st-4th, 10th, Doctorate, 5th-6th, Preschool. }
\item{education-num}{continuous}
\item{marital-status}{Married-civ-spouse, Divorced, Never-married, Separated, Widowed, Married-spouse-absent, Married-AF-spouse. }
\item{occupation}{Tech-support, Craft-repair, Other-service, Sales, Exec-managerial, Prof-specialty, Handlers-cleaners, Machine-op-inspct, Adm-clerical, Farming-fishing, Transport-moving, Priv-house-serv, Protective-serv, Armed-Forces. }
\item{relationship}{Wife, Own-child, Husband, Not-in-family, Other-relative, Unmarried.}
\item{race}{White, Asian-Pac-Islander, Amer-Indian-Eskimo, Other, Black.}
\item{sex}{Female, Male.}
\item{capital-gain}{continuous}
\item{capital-loss}{continuous}
\item{hours-per-week}{continuous}
\item{native-country}{United-States, Cambodia, England, Puerto-Rico, Canada, Germany, Outlying-US(Guam-USVI-etc), India, Japan, Greece, South, China, Cuba, Iran, Honduras, Philippines, Italy, Poland, Jamaica, Vietnam, Mexico, Portugal, Ireland, France, Dominican-Republic, Laos, Ecuador, Taiwan, Haiti, Columbia, Hungary, Guatemala, Nicaragua, Scotland, Thailand, Yugoslavia, El-Salvador, Trinadad&Tobago, Peru, Hong, Holand-Netherlands.}
\item{y}{If an individuals salary is greater than $50K/yr}
}
}
\source{
\url{https://archive.ics.uci.edu/ml/datasets/adult}
}
\usage{
adult
}
\description{
Goal is to predict whether income exceeds $50K/yr based on census data. Also known as "Census Income" dataset
}
\keyword{datasets}
