% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compas_data.R
\docType{data}
\name{compas_data}
\alias{compas_data}
\title{ProPublica's Compas recdivism data}
\format{
A data frame with 6172 rows and 17 variables:
\describe{
\item{age}{numeric}
\item{c_charge_degree}{c_charge_degree}
\item{race}{race}
\item{age_cat}{age category}
\item{score_text}{compas score text}
\item{priors_count}{number of priors}
\item{days_b_screening_arrest}{numeric}
\item{decile_score}{compas decile score}
\item{is_recid}{is_recid}
\item{two_year_recid}{two_year_recid}
\item{c_jail_in}{c_jail_in}
\item{c_jail_out}{c_jail_out}
\item{row_num}{row number}
\item{score_prob}{1 divided by the decile score}
\item{is_recid_new}{2 if recdivated and 1 if they did not. This fits the notation of the FASI paper.}
\item{race_factor}{race converted as a factor variable}
\item{sex}{male or female}
}
}
\source{
\url{https://github.com/propublica/compas-analysis}
}
\usage{
compas_data
}
\description{
ProPublica's Compas recdivism data
}
\keyword{datasets}
